refactor(onboarding): optimize onboarding container logic
- Replace export with default export for OnboardingContainer
- Use memoization for performance improvements
- Refactor handleScrollEnd and handleNext to use useCallback
- Update isLastSlide to use useMemo
- Remove unused onboarding index file
