--- src/features/onboarding/index.ts
@@ -1,2 +1,0 @@
-export { default as OnboardingScreen } from './page';
-export * from './hooks/use-onboarding-first-access';
